+++ test/admin/collections/CustomFields/index.ts
@@ -1,6 +1,9 @@
 import type { CollectionConfig } from 'payload'
 
 import { customFieldsSlug } from '../../slugs.js'
+import { AfterInput } from './AfterInput.js'
+import { BeforeInput } from './BeforeInput.js'
+import { CustomError } from './CustomError.js'
 import { FieldDescriptionComponent } from './FieldDescription/index.js'
 import { CustomSelect } from './fields/Select/index.js'
 import { CustomDescription } from './fields/Text/Description.js'
@@ -9,6 +12,21 @@ import { CustomLabel } from './fields/Text/Label.js'
 export const CustomFields: CollectionConfig = {
   slug: customFieldsSlug,
   fields: [
+    {
+      name: 'customTextField',
+      type: 'text',
+      maxLength: 100,
+      admin: {
+        components: {
+          afterInput: [AfterInput],
+          beforeInput: [BeforeInput],
+          Label: CustomLabel,
+          Description: CustomDescription,
+          Error: CustomError,
+        },
+      },
+      minLength: 3,
+    },
     {
       name: 'descriptionAsString',
       type: 'text',
@@ -29,17 +47,6 @@ export const CustomFields: CollectionConfig = {
       admin: {
         components: {
           Description: FieldDescriptionComponent,
-        },
-      },
-    },
-    {
-      name: 'customTextField',
-      type: 'text',
-      maxLength: 100,
-      admin: {
-        components: {
-          Label: CustomLabel,
-          Description: CustomDescription,
         },
       },
     },
